LibPDF: Let LineDashPattern store data as floats
We used to store them as ints, which threw away information.
In practice, files do use fractional values for these.

+++ Userland/Libraries/LibPDF/Renderer.h
@@ -37,8 +37,8 @@ enum class LineJoinStyle : u8 {
 };
 
 struct LineDashPattern {
-    Vector<int> pattern;
-    int phase;
+    Vector<float> pattern;
+    float phase;
 };
 
 enum class TextRenderingMode : u8 {
